Don't re-discover the pin modes on every state change timer event
Previously, the port register were re-read to discover the pin mode,
when that info was already kept in the digitalPinMode and analogPinMode
arrays.

This does mean that if a pin mode is changed through the Arduino or
bitlash API, that the pin change handler will miss it. But if we want to
support this case, we should probably add a separate "check pin modes"
handler that runs less often.