Introduce a new "UNSET" pin mode
This is similar to the existing "DISABLED" mode, except that the actual
pin modes aren't touched (should default to INPUT, though) and a pin can
never be set to UNSET again (unless through a reboot).

This new mode is needed for properly supporting the "make all other pins
disconnected" later.